# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateRequestBody": {
        "description": "Create a media request by tmdb id + media type.",
        "properties": {
          "media_type": {
            "enum": [
              "movie",
              "tv"
            ],
            "title": "Media Type",
            "type": "string"
          },
          "seasons": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seasons"
          },
          "tmdb_id": {
            "title": "Tmdb Id",
            "type": "integer"
          }
        },
        "required": [
          "tmdb_id",
          "media_type"
        ],
        "title": "CreateRequestBody",
        "type": "object"
      },
...
      "ErrorDetail": {
        "description": "Machine-readable error body returned by manual HTTPException paths.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KeepForeverBody": {
        "description": "``POST /api/v1/requests/{id}/keep-forever`` -- set or clear the pin.",
        "properties": {
          "keep_forever": {
            "title": "Keep Forever",
            "type": "boolean"
          }
        },
        "required": [
          "keep_forever"
        ],
        "title": "KeepForeverBody",
        "type": "object"
      },
...
      "ReportIssueBody": {
        "description": "``POST /requests/{id}/report-issue`` -- report a bad imported file (ADR-0014).\n\n``reason`` is one of the operator-choosable :class:`BlocklistReason` values\n(``failed`` is auto-only and deliberately excluded). ``season`` is REQUIRED for\na tv request (report-issue is per-season, mirroring grab) and ignored for a\nmovie. The verb blocklists the culprit release, purges the torrent + library\nfile, and synchronously re-searches (see ``correction_service.report_issue``).",
        "properties": {
          "reason": {
            "enum": [
              "bad_quality",
              "wrong_media",
              "user_reported"
            ],
            "title": "Reason",
            "type": "string"
          },
          "season": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ReportIssueBody",
        "type": "object"
      },
      "RequestListResponse": {
        "description": "A list of media requests.",
        "properties": {
          "requests": {
            "items": {
              "$ref": "#/components/schemas/RequestResponse"
            },
            "title": "Requests",
            "type": "array"
          }
        },
        "required": [
          "requests"
        ],
        "title": "RequestListResponse",
        "type": "object"
      },
      "RequestResponse": {
        "description": "A media request as returned to the client.",
        "properties": {
          "backdrop_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Backdrop Url"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_anime": {
            "default": false,
            "title": "Is Anime",
            "type": "boolean"
          },
          "keep_forever": {
            "default": false,
            "title": "Keep Forever",
            "type": "boolean"
          },
          "media_type": {
            "title": "Media Type",
            "type": "string"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "seasons": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SeasonStatus"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seasons"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tmdb_id": {
            "title": "Tmdb Id",
            "type": "integer"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "tmdb_id",
          "media_type",
          "title",
          "status"
        ],
        "title": "RequestResponse",
        "type": "object"
      },
...
      "SeasonStatus": {
        "description": "One tracked season's status, embedded in a tv ``RequestResponse``.",
        "properties": {
          "season_number": {
            "title": "Season Number",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "season_number",
          "status"
        ],
        "title": "SeasonStatus",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Api-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Plex Manager",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/requests": {
      "get": {
        "description": "List all media requests.",
        "operationId": "list_requests_endpoint_api_v1_requests_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestListResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Requests Endpoint",
        "tags": [
          "requests"
        ]
      },
      "post": {
        "description": "Create a request (or return the existing active one for this media).\n\nIf Plex is configured and the movie is already in the library, the request is\nrecorded directly as ``available`` (no needless search/grab). For a tv\nrequest, ``body.seasons`` (omitted/empty = the whole aired series) is threaded\nto ``request_service.create_request``, which tracks each named season as its\nown ``SeasonRequest`` row -- including on the dedup path, where a repeat POST\nwith a NEW season list grows the tracked set rather than being dropped.",
        "operationId": "create_request_endpoint_api_v1_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Existing matching request"
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Media not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Request Endpoint",
        "tags": [
          "requests"
        ]
      }
    },
    "/api/v1/requests/{request_id}": {
      "get": {
        "description": "Return a single media request, or 404.",
        "operationId": "get_request_endpoint_api_v1_requests__request_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Request not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Get Request Endpoint",
        "tags": [
          "requests"
        ]
      }
    },
    "/api/v1/requests/{request_id}/cancel": {
      "post": {
        "description": "Cancel a not-yet-imported request (ADR-0014): drop active torrent(s), settle.\n\nRemoves any active torrent(s) WITH data (best-effort) and flips the request\n(and every tracked season, for tv) to the settled ``cancelled`` status; the\nrow is kept for history and nothing is re-grabbed. A request past the\nnot-yet-imported stage is refused (409 ``not_cancellable``) -- use report-issue\nto redo an imported title instead.\n\nqBittorrent is resolved OPTIONALLY (``get_qbittorrent_optional``): a cancel for a\n``pending``/``searching``/``no_acceptable_release`` request with NO active download\nrows is a pure DB settle that never touches the client, so it still works on an\ninstall with qBittorrent unconfigured. When there ARE active torrents to remove but\nthe client is unconfigured, the service refuses up front (409\n``service_not_configured``) rather than silently leaking a seeding torrent.",
        "operationId": "cancel_request_endpoint_api_v1_requests__request_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Cancel Request Endpoint",
        "tags": [
          "requests"
        ]
      }
    },
    "/api/v1/requests/{request_id}/keep-forever": {
      "post": {
        "description": "Set or clear the operator's \"keep forever\" pin (ADR-0012).\n\nThe north-star #1 correction path for \"don't let the eviction sweep touch\nthis one\": a pinned title (or, for a show, every one of its seasons -- the\npin lives on the parent) is never selected by ``domain/eviction.py``\nregardless of watch state or disk pressure. A 404 for an unknown id, never\na silent no-op.",
        "operationId": "keep_forever_endpoint_api_v1_requests__request_id__keep_forever_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeepForeverBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Keep Forever Endpoint",
        "tags": [
          "requests"
        ]
      }
    },
    "/api/v1/requests/{request_id}/report-issue": {
      "post": {
        "description": "Report a bad imported/available movie or TV season (ADR-0014).\n\nBlocklists the culprit release, removes its torrent + the library file, and\nsynchronously re-searches for a DIFFERENT release (the honest\n``no_acceptable_release`` park if nothing is acceptable). Requires Plex +\nqBittorrent + Prowlarr configured (their deps 409 ``service_not_configured``\notherwise). The correction-without-a-terminal button for \"this file is bad\".",
        "operationId": "report_issue_endpoint_api_v1_requests__request_id__report_issue_post",
        "parameters": [
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportIssueBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Report Issue Endpoint",
        "tags": [
          "requests"
        ]
      }
    },
...
  }
}